+++ app/src/main/java/com/example/buggyweather/home/HomeModule.kt
@@ -1,8 +1,8 @@
 package com.example.buggyweather.home
 
 import androidx.arch.core.util.Function
-import com.example.buggyweather.base.KoinModule
-import com.example.buggyweather.base.UseCase
+import com.example.buggyweather.core.base.KoinModule
+import com.example.buggyweather.core.base.UseCase
 import com.example.buggyweather.domain.CurrentWeather
 import com.example.buggyweather.domain.MeasuringUnits
 import com.example.buggyweather.home.presenter.HomeViewModel
