+++ app/src/main/java/com/example/buggyweather/home/presenter/HomeFragment.kt
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
 import com.example.buggyweather.R
-import com.example.buggyweather.base.BaseFragment
+import com.example.buggyweather.core.base.BaseFragment
 import com.example.buggyweather.databinding.FragmentHomeBinding
 import com.example.buggyweather.domain.Coordinate
 import com.example.buggyweather.domain.CurrentWeather
