+++ app/src/main/java/com/example/buggyweather/home/presenter/HomeViewModel.kt
@@ -3,8 +3,8 @@ package com.example.buggyweather.home.presenter
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
-import com.example.buggyweather.base.BaseViewModel
-import com.example.buggyweather.base.UseCase
+import com.example.buggyweather.core.base.BaseViewModel
+import com.example.buggyweather.core.base.UseCase
 import com.example.buggyweather.domain.CurrentWeather
 import com.example.buggyweather.domain.MeasuringUnits
 import com.example.buggyweather.network.exception.BadRequestException
