+++ app/src/main/java/com/example/buggyweather/home/usecase/GetCurrentWeatherUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.buggyweather.home.usecase
 
-import com.example.buggyweather.base.UseCase
+import com.example.buggyweather.core.base.UseCase
 import com.example.buggyweather.domain.CurrentWeather
 import com.example.buggyweather.domain.MeasuringUnits
 import com.example.buggyweather.home.repository.CurrentWeatherDataSource
