+++ app/src/main/java/com/example/buggyweather/whole/WholeDayModule.kt
@@ -1,8 +1,8 @@
 package com.example.buggyweather.whole
 
 import androidx.arch.core.util.Function
-import com.example.buggyweather.base.KoinModule
-import com.example.buggyweather.base.UseCase
+import com.example.buggyweather.core.base.KoinModule
+import com.example.buggyweather.core.base.UseCase
 import com.example.buggyweather.domain.ForecastRequest
 import com.example.buggyweather.domain.WholeDayWeather
 import com.example.buggyweather.whole.presenter.WholeDayViewModel
