+++ app/src/test/java/com/example/buggyweather/main/repository/LastCityRepositoryTest.kt
@@ -22,7 +22,7 @@ class LastCityRepositoryTest : BaseTest() {
 
 	@Before
 	fun setUp() {
-		MockitoAnnotations.initMocks(this)
+		MockitoAnnotations.openMocks(this)
 		repository = LastCityRepository(sharedPref)
 	}
