+++ app/src/test/java/com/example/buggyweather/main/usecase/GetMeasuringUnitsUseCaseTest.kt
@@ -1,6 +1,6 @@
 package com.example.buggyweather.main.usecase
 
-import com.example.buggyweather.base.UseCase
+import com.example.buggyweather.core.base.UseCase
 import com.example.buggyweather.domain.MeasuringUnits
 import com.example.buggyweather.helper.BaseTest
 import com.example.buggyweather.main.repository.MeasuringUnitsDataSource
@@ -24,7 +24,7 @@ class GetMeasuringUnitsUseCaseTest : BaseTest() {
 
 	@Before
 	fun setUp() {
-		MockitoAnnotations.initMocks(this)
+		MockitoAnnotations.openMocks(this)
 		useCase = GetMeasuringUnitsUseCase(dataSource)
 	}
