+++ app/src/test/java/com/example/buggyweather/main/usecase/SaveMeasuringUnitUseCaseTest.kt
@@ -1,6 +1,6 @@
 package com.example.buggyweather.main.usecase
 
-import com.example.buggyweather.base.UseCase
+import com.example.buggyweather.core.base.UseCase
 import com.example.buggyweather.domain.MeasuringUnits
 import com.example.buggyweather.helper.BaseTest
 import com.example.buggyweather.main.repository.MeasuringUnitsDataSource
@@ -21,7 +21,7 @@ class SaveMeasuringUnitUseCaseTest : BaseTest() {
 
 	@Before
 	fun setUp() {
-		MockitoAnnotations.initMocks(this)
+		MockitoAnnotations.openMocks(this)
 		useCase = SaveMeasuringUnitsUseCase(dataSource)
 	}
