+++ common.gradle
@@ -12,7 +12,7 @@ ext {
     tests = [
         "androidx.arch.core:core-testing:2.1.0",
         'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.4.2',
-        "org.mockito:mockito-core:3.3.1",
+        "org.mockito:mockito-core:3.12.4",
         "com.nhaarman.mockitokotlin2:mockito-kotlin:2.0.0"
     ]
