+++ core/src/main/java/com/example/buggyweather/core/base/BaseViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.buggyweather.base
+package com.example.buggyweather.core.base
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
