+++ core/src/main/java/com/example/buggyweather/core/base/UseCase.kt
@@ -1,4 +1,4 @@
-package com.example.buggyweather.base
+package com.example.buggyweather.core.base
 
 abstract class UseCase<REQUEST, RESPONSE> {
